backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
dj-database-url==2.1.0
Django==4.2.30
django-axes==8.3.1
django-celery-beat==2.7.0
django-celery-results==2.6.0
django-ckeditor-5==0.2.12
django-cors-headers==4.6.0
django-js-asset==2.2.0
django-rest-passwordreset==1.5.0
django-timezone-field==7.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.27.2
setuptools<81
gunicorn==23.0.0
idna==3.10
jsonfield==3.1.0
kombu==5.4.2
MarkupSafe==3.0.2
packaging==24.2
pillow==12.2.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
cryptography>=41.0.0
PyJWT==2.12.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
pytz==2024.2
redis==5.2.1
requests==2.33.1
openai>=1.35.0
pgvector>=0.3.0
six==1.16.0
sqlparse==0.5.5
stripe==12.0.0
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
sentry-sdk[django]==2.19.0
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
django-anymail[resend]
groq>=0.12.0
